chore: relocate ai slice modal
