+++ packages/slice-machine/src/features/customTypes/GenerateSliceWithAiModal/SliceCard.tsx
